refactor(native): isolate Objective-C compilation
