+++ proton/internal/native/cef_loader/moon.pkg
@@ -2,12 +2,7 @@ supported_targets = "native"
 
 options(
   "is-main": false,
-  "native-stub": [ "loader_stub.mm" ],
-  link: {
-    "native": {
-      "stub-cc": "${build.PROTON_CEF_LOADER_CC}",
-      "stub-cc-flags": "${build.PROTON_CEF_LOADER_CC_FLAGS}",
-    },
-  },
+  "native-stub": [ "loader_stub.objcxx.c" ],
+  link: { "native": { "stub-cc-flags": "${build.PROTON_CEF_LOADER_CC_FLAGS}" } },
   targets: { "loader.mbt": [ "native" ] },
 )
