+++ proton/internal/native/ffi/ffi.mbt
@@ -1,5 +1,6 @@
 ///|
 pub fn ensure_cef_loader_linked() -> Unit {
+  @mac_engine.ensure_linked()
   @cef_loader.ensure_linked()
 }
