+++ proton/internal/native/ffi_mac/mac_client.objc.c
@@ -2,14 +2,14 @@
 
 #include "mac_internal.h"
 
-#include "../../proton_config.h"
-#include "../../proton_event.h"
-#include "../../proton_json.h"
-#include "../cef_common/bridge_lifecycle.h"
-#include "../cef_common/bridge_json.h"
-#include "../cef_common/browser_session.h"
-#include "../cef_common/bridge_renderer.h"
-#include "../cef_common/message.h"
+#include "../ffi/src/proton_config.h"
+#include "../ffi/src/proton_event.h"
+#include "../ffi/src/proton_json.h"
+#include "../ffi/src/engine/cef_common/bridge_lifecycle.h"
+#include "../ffi/src/engine/cef_common/bridge_json.h"
+#include "../ffi/src/engine/cef_common/browser_session.h"
+#include "../ffi/src/engine/cef_common/bridge_renderer.h"
+#include "../ffi/src/engine/cef_common/message.h"
 #define PROTON_ENGINE_REF_INCREMENT(refs) \
   atomic_fetch_add_explicit(&(refs)->refs, 1, memory_order_relaxed)
 #define PROTON_ENGINE_REF_DECREMENT(refs) \
@@ -17,14 +17,14 @@
 #define PROTON_ENGINE_REF_LOAD(refs) \
   atomic_load_explicit(&(refs)->refs, memory_order_acquire)
 #define PROTON_ENGINE_REF_STORE(refs, value) atomic_store(&(refs)->refs, value)
-#include "../cef_common/ref_count.h"
+#include "../ffi/src/engine/cef_common/ref_count.h"
 #undef PROTON_ENGINE_REF_INCREMENT
 #undef PROTON_ENGINE_REF_DECREMENT
 #undef PROTON_ENGINE_REF_LOAD
 #undef PROTON_ENGINE_REF_STORE
-#include "../cef_common/scheme.h"
-#include "../cef_common/strings.h"
-#include "../cef_common/view_events.h"
+#include "../ffi/src/engine/cef_common/scheme.h"
+#include "../ffi/src/engine/cef_common/strings.h"
+#include "../ffi/src/engine/cef_common/view_events.h"
 #include "mac_dialog.h"
 #include "mac_launch_input.h"
