+++ proton/internal/native/ffi_mac/mac_view.objc.c
@@ -2,11 +2,11 @@
 
 #include "mac_internal.h"
 
-#include "../../proton_config.h"
-#include "../../proton_event.h"
-#include "../../proton_json.h"
-#include "../cef_common/browser_session.h"
-#include "../cef_common/message.h"
+#include "../ffi/src/proton_config.h"
+#include "../ffi/src/proton_event.h"
+#include "../ffi/src/proton_json.h"
+#include "../ffi/src/engine/cef_common/browser_session.h"
+#include "../ffi/src/engine/cef_common/message.h"
 #define PROTON_ENGINE_REF_INCREMENT(refs) \
   atomic_fetch_add_explicit(&(refs)->refs, 1, memory_order_relaxed)
 #define PROTON_ENGINE_REF_DECREMENT(refs) \
@@ -14,14 +14,14 @@
 #define PROTON_ENGINE_REF_LOAD(refs) \
   atomic_load_explicit(&(refs)->refs, memory_order_acquire)
 #define PROTON_ENGINE_REF_STORE(refs, value) atomic_store(&(refs)->refs, value)
-#include "../cef_common/ref_count.h"
+#include "../ffi/src/engine/cef_common/ref_count.h"
 #undef PROTON_ENGINE_REF_INCREMENT
 #undef PROTON_ENGINE_REF_DECREMENT
 #undef PROTON_ENGINE_REF_LOAD
 #undef PROTON_ENGINE_REF_STORE
-#include "../cef_common/scheme.h"
-#include "../cef_common/strings.h"
-#include "../cef_common/view_events.h"
+#include "../ffi/src/engine/cef_common/scheme.h"
+#include "../ffi/src/engine/cef_common/strings.h"
+#include "../ffi/src/engine/cef_common/view_events.h"
 
 #include "include/capi/cef_browser_capi.h"
 #include "include/capi/cef_frame_capi.h"
