+++ proton/internal/native/ffi_mac/mac_window.objc.c
@@ -2,15 +2,15 @@
 
 #include "mac_internal.h"
 
-#include "../../proton_config.h"
-#include "../../proton_event.h"
-#include "../cef_common/bridge_lifecycle.h"
-#include "../cef_common/browser_session.h"
-#include "../cef_common/bridge_renderer.h"
-#include "../cef_common/message.h"
-#include "../cef_common/scheme.h"
-#include "../cef_common/strings.h"
-#include "../cef_common/view_events.h"
+#include "../ffi/src/proton_config.h"
+#include "../ffi/src/proton_event.h"
+#include "../ffi/src/engine/cef_common/bridge_lifecycle.h"
+#include "../ffi/src/engine/cef_common/browser_session.h"
+#include "../ffi/src/engine/cef_common/bridge_renderer.h"
+#include "../ffi/src/engine/cef_common/message.h"
+#include "../ffi/src/engine/cef_common/scheme.h"
+#include "../ffi/src/engine/cef_common/strings.h"
+#include "../ffi/src/engine/cef_common/view_events.h"
 #include "mac_dialog.h"
 #include "mac_launch_input.h"
 #include "mac_menu.h"
